Move ScheduledMessage model from chat app to events app and drop the ScheduledMessageConfig model
