+++ apps/events/tasks.py
@@ -1,8 +1,9 @@
 import logging
 
 from celery.app import shared_task
+from django.db.models import functions
 
-from apps.events.models import StaticTrigger, TimeoutTrigger
+from apps.events.models import ScheduledMessage, StaticTrigger, TimeoutTrigger
 from apps.experiments.models import ExperimentSession
 
 logger = logging.getLogger(__name__)
@@ -49,3 +50,17 @@ def fire_trigger(trigger_id, session_id):
     session = ExperimentSession.objects.get(id=session_id)
     triggered = trigger.fire(session)
     return triggered
+
+
+def _get_messages_to_fire():
+    return ScheduledMessage.objects.filter(is_complete=False, next_trigger_date__lte=functions.Now())
+
+
+@shared_task()
+def poll_scheduled_messages():
+    """Polls scheduled messages and triggers those that are due. After triggering, it updates the database with the
+    new trigger details for each message."""
+
+    messages = _get_messages_to_fire()
+    for message in messages:
+        message.safe_trigger()
